aarch64: fix tcsh fd-16 logout — register_t must be the machine word
The disk login chain runs the userdb shell /bin/tcsh, which immediately
logged out: it relocates its stdio onto high descriptors (initdesc:
dcopy(0,FSHIN=16) etc.) then reads FSHIN, but fd 16 was never created so
the read returned -1 -> logout.  Three layered defects, root cause last:

1. register_t was 'typedef int' for all arches.  On aarch64 (LP64) a
   trapframe register slot is 64-bit, but ksyscall_dispatch casts the
   saved-arg array to (register_t *) and indexes it, so with a 32-bit
   register_t every table-dispatched syscall read its 2nd and later args
   from the wrong half of a register.  dup2(old,new) saw new=0, so
   dcopy(0,16) became dup2(0,0) (a no-op) and fd 16 was never populated.
   Every special-cased syscall (read/write/open/mmap...) worked only
   because it bypassed the table and read args[] as 64-bit words directly.
   Fix: typedef long register_t — 4 bytes on i386 ILP32 (byte-identical),
   8 bytes on aarch64 LP64.  This repairs the generic dispatch for ALL
   multi-arg table syscalls on aarch64, not just dup2.

2. aarch64_fork never copied the parent's open-file table, so the forked
   child that exec'd tcsh started with schedNewTask's blank placeholder
   fds (no fileops) instead of login's console fds.  Add fd-table
   inheritance mirroring i386 fork, plus cwd/pgrp/sid/ct_tty/uid/gid.

3. aarch64_console_setup_fds used falloc (lowest free slot), which skipped
   the placeholders schedNewTask seeds at 0-2 and put the console at 3/4/5,
   leaving 0/1/2 blank.  Populate slots 0/1/2 in place.

Also: aarch64 sys_getcwd returned 0 on success, but musl's getcwd() treats
0 as ENOENT (it wants the path length).  Return strlen+1 like the shared
i386 sys_getcwd; this clears tcsh's spurious 'No such file or directory'.

QEMU-verified (HVF): root/user -> tcsh interactive prompt, echo + ls run,
and udptest does a full DNS round-trip over virtio-net/SLIRP (socket ->
sendto 8.8.8.8:53 -> poll -> recvfrom, ancount=1) — userland sockets (#2)
validated end-to-end.  i386 builds + boots to a working tcsh unchanged.

+++ sys/include/sys/types.h
@@ -55,7 +55,17 @@ typedef unsigned int   u_int;
 typedef unsigned long  u_long;
 
 typedef int pidType;
-typedef int register_t;
+/*
+ * register_t is the natural machine word (FreeBSD semantics): 32-bit on the
+ * i386 ILP32 target, 64-bit on the aarch64 LP64 target.  `long` is exactly
+ * this width on both (4 bytes ILP32, 8 bytes LP64), so i386 is byte-identical.
+ * The syscall argument structs (PADL_/PADR_ padding) and the table dispatcher
+ * (ksyscall_dispatch) size their arguments in units of register_t and index
+ * the saved trapframe register slots as register_t words — so this MUST match
+ * the width of one trapframe register, or every table-dispatched syscall reads
+ * its second and later arguments from the wrong half of a register.
+ */
+typedef long register_t;
 
 typedef int pid_t;
 typedef int size_t; /* standard */
